refactor: step4 클래스 이름 변경
