+++ src/main/java/lotto/domain/WinningResult.java
@@ -3,17 +3,17 @@ package lotto.domain;
 import java.util.Collections;
 import java.util.Map;
 
-public class LottoTicketsWinInformation {
+public class WinningResult {
     private final Map<LottoRank, Integer> winningStaticsMap;
     private final int lottoTicketsPrice;
 
-    private LottoTicketsWinInformation(Map<LottoRank, Integer> winningStaticsMap, int lottoTicketsPrice) {
+    private WinningResult(Map<LottoRank, Integer> winningStaticsMap, int lottoTicketsPrice) {
         this.winningStaticsMap = winningStaticsMap;
         this.lottoTicketsPrice = lottoTicketsPrice;
     }
-    
-    public static LottoTicketsWinInformation from(Map<LottoRank, Integer> winningStaticsMap, int price) {
-        return new LottoTicketsWinInformation(winningStaticsMap, price);
+
+    public static WinningResult from(Map<LottoRank, Integer> winningStaticsMap, int price) {
+        return new WinningResult(winningStaticsMap, price);
     }
 
     public Map<LottoRank, Integer> winningStatics() {
